Fixed ctrl-click behaviour on navbar
Previously, it did not open in a new tab.
Now it does.

However, this introduces a regression that was fixed in 97a54bd
Now if a user clicks on 'Create Tournament' when at /create-tournament,
and then tries to press back, the user will still be at
/create-tournament until the user presses back again.

Personally I feel as if browsing back through the history is less
of an issue compared to non-standard ctrl-click behaviour.

The issue with react-router Link pushing the same path multiple times
can be tracked in:
https://github.com/ReactTraining/react-router/issues/5039
https://github.com/ReactTraining/history/issues/470