Make default multipleDanceScoringRule "average"
While not the perfect fix, this ensures that for any round
that does not explicitly set this configuration (e.g. any round
with only a single dance per person, which should be any round
but the final), a person that dances multiple dances will have
his/her scored averaged over the dances.

Closes #134